requirements.txt
Fireworks>=1.4.3
PyYAML
numpy>=1.12.1
pandas>=0.20.1
docopt
dill
scipy
statsmodels
